requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-filter==22.1
django-guardian==2.4.0
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
importlib-metadata==4.12.0
Markdown==3.4.1
phonenumbers==8.12.53
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
zipp==3.8.1
